fix: rename cmd cp to copy
