+++ internal/component/copy/cp.go
@@ -1,4 +1,4 @@
-package cp
+package copy
 
 import (
 	"errors"
